+++ src/fhe_attack_replay/attacks/__init__.py
@@ -1,7 +1,7 @@
 # Copyright 2026 Vaultbytes (Bader Issaei)
 # SPDX-License-Identifier: Apache-2.0
 
-from fhe_attack_replay.attacks.base import Attack, AttackResult, AttackStatus
+from fhe_attack_replay.attacks.base import Attack, AttackIntent, AttackResult, AttackStatus
 from fhe_attack_replay.attacks.cheon_2024_127 import Cheon2024_127
 from fhe_attack_replay.attacks.eprint_2025_867 import Eprint2025_867
 from fhe_attack_replay.attacks.glitchfhe_usenix25 import GlitchFHE_USENIX25
@@ -10,6 +10,7 @@ from fhe_attack_replay.attacks.reveal_2023_1128 import RevEAL_2023_1128
 
 __all__ = [
     "Attack",
+    "AttackIntent",
     "AttackResult",
     "AttackStatus",
     "Cheon2024_127",